Move SearchTextField to the correct package
